requirements.txt
-i https://pypi.org/simple
aniso8601==9.0.1
asgiref==3.7.2; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
django==4.2.8
django-annoying==0.10.6
django-cors-headers==4.3.1
django-crontab==0.7.1
django-filter==22.1; python_version >= '3.7'
djangorestframework==3.14.0
graphene==3.3
graphene-django==3.2.0
graphene-django-extras==1.0.0
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
mysqlclient==2.1.1
pillow==10.1.0
promise==2.3
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.9.0; python_version < '3.11'
